+++ NEWS.md
@@ -1,4 +1,8 @@
-# geotargets 0.0.0.9000 (development version)
+# geotargets (development version)
+
+* Created `tar_stars()` and `tar_stars_proxy()` that create `stars` and `stars_proxy` objects, respectively.
+
+# geotargets 0.1.0 (14 May 2024)
 
 * Created `tar_terra_rast()` and `tar_terra_vect()` for targets that create `SpatRaster` and `SpatVector` objects, respectively
 * Created `tar_terra_sprc()` that creates a `SpatRasterCollection` object.